+++ code.rkt
@@ -10,6 +10,7 @@
          tex-code
          latex-code
          objc-code
+         pascal-code
          plist-code
          csv-code
          html-code
@@ -33,6 +34,7 @@
          texblock
          latexblock
          objcblock
+         pascalblock
          plistblock
          csvblock
          htmlblock
@@ -56,6 +58,7 @@
          texblock0
          latexblock0
          objcblock0
+         pascalblock0
          plistblock0
          csvblock0
          htmlblock0
@@ -90,6 +93,7 @@
 (define-syntax-rule (tex-code . rest) (lang:tex-code . rest))
 (define-syntax-rule (latex-code . rest) (lang:latex-code . rest))
 (define-syntax-rule (objc-code . rest) (lang:objc-code . rest))
+(define-syntax-rule (pascal-code . rest) (lang:pascal-code . rest))
 (define-syntax-rule (plist-code . rest) (lang:plist-code . rest))
 (define-syntax-rule (csv-code . rest) (lang:csv-code . rest))
 (define-syntax-rule (html-code . rest) (lang:html-code . rest))
@@ -113,6 +117,7 @@
 (define-syntax-rule (texblock . rest) (lang:texblock . rest))
 (define-syntax-rule (latexblock . rest) (lang:latexblock . rest))
 (define-syntax-rule (objcblock . rest) (lang:objcblock . rest))
+(define-syntax-rule (pascalblock . rest) (lang:pascalblock . rest))
 (define-syntax-rule (plistblock . rest) (lang:plistblock . rest))
 (define-syntax-rule (csvblock . rest) (lang:csvblock . rest))
 (define-syntax-rule (htmlblock . rest) (lang:htmlblock . rest))
@@ -136,6 +141,7 @@
 (define-syntax-rule (texblock0 . rest) (lang:texblock0 . rest))
 (define-syntax-rule (latexblock0 . rest) (lang:latexblock0 . rest))
 (define-syntax-rule (objcblock0 . rest) (lang:objcblock0 . rest))
+(define-syntax-rule (pascalblock0 . rest) (lang:pascalblock0 . rest))
 (define-syntax-rule (plistblock0 . rest) (lang:plistblock0 . rest))
 (define-syntax-rule (csvblock0 . rest) (lang:csvblock0 . rest))
 (define-syntax-rule (htmlblock0 . rest) (lang:htmlblock0 . rest))